% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autismDeNovos.doc.R
\docType{data}
\name{autismDeNovos}
\alias{autismDeNovos}
\title{de novo variants found in 1,078 autism trios}
\format{A data frame with 1096 obs of 2 variables:
\describe{
  \item{gene}{Gene symbol of gene containing de novo variant}
  \item{class}{Functional class of variant:
    "syn" = synonymous,
    "mis" = missense,
    "non" = nonsense,
    "splice" = canonical splice site,
    "frameshift" = frameshift indel}
}}
\description{
de novo variants found in 1,078 autism trios, published in Nature Genetics(http://www.nature.com/doifinder/10.1038/ng.3050)
}
\author{
James Ware
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4222185/}
}
\keyword{datasets}

