% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/denovolyzeMultiHits.R
\name{denovolyzeMultiHits}
\alias{denovolyzeMultiHits}
\title{Determine significance of genes with multiple \emph{de novos}}
\usage{
denovolyzeMultiHits(dnm.genes, dnm.classes, nsamples, nperms = 100,
  include.gene = "all", include.class = c("syn", "mis", "lof", "prot",
  "all"), expectedDNMs = "actual", gene.id = "hgncID", probTable = NULL,
  mis_filter = NULL)
}
\arguments{
\item{dnm.genes}{A vector of genes containing de novo variants.}

\item{dnm.classes}{A vector of classes of de novo variants.  Supported
classes are "syn", "mis", "mis_cons", "mis_dam", "mis_oth","non" or
"stoploss","startloss","splice", "frameshift" and "lof".}

\item{nsamples}{Number of individuals considered in de novo analysis.}

\item{nperms}{Number of permutations}

\item{include.gene}{Genes to include in analysis. "all" or a vector of gene
names.}

\item{include.class}{Which variant classes are tabulated in output}

\item{expectedDNMs}{Select whether expected number of multihits is determined
  by expected total de novos, or actual total}

\item{gene.id}{Gene identifier used. Currently only hgncID. (refseqID may
work)}

\item{probTable}{Probability table. A user-defined table of probabilities can be
provided here, to replace the probability table included in the package.}

\item{mis_filter}{If the user-specified probability table contains
  probabilities for a sub-category of missense variants (e.g. predicted to be
  damaging by an in silico algorithm), thta column should be called
  mis_filter, or the alternative name should be specified here.}
}
\value{
Returns a data.frame
}
\description{
Are there more genes containing >1 \emph{de novos} than expected?
}
\details{
See vignette (denovostats_intro) for more information.
}
\examples{
denovolyzeMultiHits(dnm.genes=autismDeNovos$gene,
                    dnm.classes=autismDeNovos$dnmClass,
                    nsamples=1078)
}
\keyword{keywords}

