% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fmrpGenes.doc.R
\docType{data}
\name{fmrpGenes}
\alias{fmrpGenes}
\title{FMRP genes}
\format{A vector of gene symbols}
\description{
837 genes found to interact with the fragile X mental retardation protein (FMRP)
}
\author{
James Ware
}
\references{
\url{http://www.nature.com/ng/journal/v46/n9/full/ng.3050.html}

\url{http://dx.doi.org/10.1016/j.cell.2011.06.013}
}
\keyword{datasets}

