% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parseInput}
\alias{parseInput}
\title{Checks input for errors}
\usage{
parseInput(dnm.genes = dnm.genes, dnm.classes = dnm.genes,
  nsamples = nsamples, group.by = group.by, include.gene = include.gene,
  include.class = include.class, gene.id = gene.id, signif.p = signif.p,
  round.expected = round.expected)
}
\value{
tbc   , with or without error message
}
\description{
An internal function to check inputs
}

