% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PermuteMultiHits}
\alias{PermuteMultiHits}
\title{Permutes x variants across a genelist, and counts genes with multiple hits}
\usage{
PermuteMultiHits(x, y, nperms = 100, class = "lof", gene.id = "hgncID",
  include.gene = "all")
}
\arguments{
\item{x}{Total number of de novo variants observed in dataset}

\item{y}{Number of genes with >1 de novo variant (of class "class") in the population}

\item{nperms}{Number permutations}

\item{class}{In c("lof","mis","syn","prot")}

\item{gene.id}{Gene identifier used. Currently only hgnc.id}

\item{include}{Genes to include in analysis. "all" or a vector of gene names}
}
\value{
Returns a named vector of 5 values
}
\description{
An internal function called by denovolyzeMultiHits
}

