% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewProbabilityTable.R
\name{viewProbabilityTable}
\alias{viewProbabilityTable}
\title{Displays underlying \emph{de novo} probability tables}
\usage{
viewProbabilityTable(format = "wide")
}
\arguments{
\item{format}{option to display table in wide format (default; one line per gene), or long format}
}
\description{
Tabulates probability of \emph{de novo} variant for each protein-coding variant class, for each gene.  Values are probability of a \emph{de novo} variant per chromosome per generation.  i.e. expected number of de novos for a given gene/class = \eqn{p * 2 * nsamples}.
}

