% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/denovolyze.R
\name{denovolyze}
\alias{denovolyze}
\alias{denovolyzeByClass}
\alias{denovolyzeByGene}
\title{Evaluates burden of \emph{de novo} variation against expectation}
\usage{
denovolyze(genes, classes, nsamples, group.by = "class",
  include.gene = "all", include.class = c("syn", "mis", "mis_filter",
  "mis_other", "mis", "non", "stoploss", "startgain", "splice", "frameshift",
  "lof", "prot", "all", "prot_dam"), gene.id = "geneID", signif.p = 3,
  round.expected = 1, probTable = NULL, mis_filter = NULL)

denovolyzeByClass(genes, classes, nsamples, group.by = "class",
  include.gene = "all", include.class = c("syn", "mis", "lof", "prot",
  "all"), gene.id = "geneID", signif.p = 3, round.expected = 1,
  probTable = NULL)

denovolyzeByGene(genes, classes, nsamples, group.by = "gene",
  include.gene = "all", include.class = c("lof", "prot"),
  gene.id = "geneID", signif.p = 3, round.expected = 1,
  probTable = NULL)
}
\arguments{
\item{genes}{A vector of genes containing de novo variants.}

\item{classes}{A vector of classes of de novo variants.  Supported
classes are "syn", "mis", "mis_cons", "mis_dam", "mis_oth","non" or
"stoploss","startloss","splice", "frameshift" and "lof".}

\item{nsamples}{Number of individuals considered in de novo analysis.}

\item{group.by}{Results can be tabulated by gene, or by variant class}

\item{include.gene}{Genes to include in analysis. "all" or a vector of gene
names.}

\item{include.class}{Which variant classes are tabulated in output}

\item{gene.id}{Gene identifier used. Currently only geneID. (refseqID may
work)}

\item{signif.p}{Number of sig figs used to round p values in output.}

\item{round.expected}{Number of decimal places used to round expected burdens
in output.}

\item{probTable}{Probability table. A user-defined table of probabilities can be
provided here, to replace the probability table included in the package.}

\item{mis_filter}{If the user-specified probability table contains
  probabilities for a sub-category of missense variants (e.g. predicted to be
  damaging by an in silico algorithm), thta column should be called
  mis_filter, or the alternative name should be specified here.}
}
\value{
Returns a data frame
}
\description{
Determines whether the test population carry more \emph{de novo} variants
than expected. Variants may be grouped by variant class (e.g. are there more
LOF variants than expected, across the whole dataset?), or by gene (are there
more variants of a given class in SCN2A?).
}
\details{
Analyses can be restricted to a subset of genes, and/or a subset of variant
classes

See vignette("denovolyzeR_intro") for more information.
}
\section{Functions}{
\itemize{
\item \code{denovolyzeByClass}: denovolyzeByClass

\item \code{denovolyzeByGene}: denovolyzeByGene
}}
\examples{
### denovolyze

denovolyze(genes=autismDeNovos$gene,
           classes=autismDeNovos$dnmClass,
           nsamples=1078)

### denovolyzeByClass

denovolyzeByClass(genes=autismDeNovos$gene,
                  classes=autismDeNovos$dnmClass,
                  nsamples=1078)

# this convenience function is identical to:

denovolyze(genes=autismDeNovos$gene,
           classes=autismDeNovos$dnmClass,
           nsamples=1078,
           group.by="class",
           include.class=c("syn","mis","lof","prot","all"),
           include.gene="all"
           )

### denovolyzeByGene

denovolyzeByGene(genes=autismDeNovos$gene,
                 classes=autismDeNovos$dnmClass,
                 nsamples=1078)

# this is identical to:

denovolyze(genes=autismDeNovos$gene,
           classes=autismDeNovos$dnmClass,
           nsamples=1078,
           group.by="gene",
           include.class=c("lof","prot"),
           include.gene="all"
           )
}

