% Generated by roxygen2 (4.0.2): do not edit by hand
\name{denovolyze}
\alias{denovolyze}
\title{Evaluates burden of \emph{de novo} variation against expectation}
\usage{
denovolyze(dnm.genes, dnm.classes, nsamples, group.by = "class",
  include.gene = "all", include.class = c("syn", "mis", "non", "splice",
  "frameshift", "lof", "prot", "all"), gene.id = "hgncID", signif.p = 3,
  round.expected = 1)
}
\arguments{
\item{dnm.genes}{A vector of genes containing de novo variants.}

\item{dnm.classes}{A vector of classes of de novo variants.  Supported
classes are "syn", "mis", "non", "splice", "frameshift" and "lof".}

\item{nsamples}{Number of individuals considered in de novo analysis.}

\item{group.by}{Results can be tabulated by gene, or by variant class}

\item{include.gene}{Genes to include in analysis. "all" or a vector of gene names.}

\item{include.class}{Which variant classes are tabulated in output}

\item{gene.id}{Gene identifier used. Currently only hgncID. (refseqID may work)}

\item{signif.p}{Number of sig figs used to round p values in output.}

\item{round.expected}{Number of decimal places used to round expected burdens
  in output.}
}
\value{
Returns a data frame
}
\description{
Determines whether the test population carry more \emph{de novo} variants
than expected. Variants may be grouped by variant class (e.g. are there more
LOF variants than expected, across the whole dataset?), or by gene (are there
more variants of a given class in SCN2A?).
}
\details{
Analyses can be restricted to a subset of genes, and/or a subset of variant classes

See vignette (denovolyeR intro) for more information.
}
\examples{
denovolyzeByClass(dnm.genes=autismDeNovos$gene,
                  dnm.classes=autismDeNovos$dnmClass,
                  nsamples=1078)

# this convenience function is identical to:

denovolyze(dnm.genes=autismDeNovos$gene,
           dnm.classes=autismDeNovos$dnmClass,
           nsamples=1078,
           group.by="class",
           include.class=c("syn","mis","lof","prot","all"),
           include.gene="all"
           )

##
denovolyzeByGene(dnm.genes=autismDeNovos$gene,
                 dnm.classes=autismDeNovos$dnmClass,
                 nsamples=1078)

# this is identical to:

denovolyze(dnm.genes=autismDeNovos$gene,
           dnm.classes=autismDeNovos$dnmClass,
           nsamples=1078,
           group.by="gene",
           include.class=c("lof","prot"),
           include.gene="all"
           )
}

