% Generated by roxygen2 (4.0.2): do not edit by hand
\name{denovolyzeMultiHits}
\alias{denovolyzeMultiHits}
\title{Determine significance of genes with multiple \emph{de novos}}
\usage{
denovolyzeMultiHits(dnm.genes, dnm.classes, nsamples, nperms = 100,
  include.gene = "all", include.class = c("syn", "mis", "lof", "prot",
  "all"), expectedDNMs = "actual", gene.id = "hgncID")
}
\arguments{
\item{dnm.genes}{A vector of genes containing de novo variants}

\item{dnm.classes}{A vector of classes of de novo variants.  Supported
classes are "syn", "mis", "lof", "prot"}

\item{nsamples}{Number of individuals considered in de novo analysis#'}

\item{nperms}{Number of permutations}

\item{include.class}{variant classes to tabulate in output.  Valid classes =
"syn","mis","non","splice","frameshift","lof","prot","all".}

\item{expectedDNMs}{Select whether expected number of multihits is determined
by expected total de novos, or actual total}

\item{gene.id}{Gene identifier used. Currently only hgnc.id}

\item{include}{Genes to include in analysis. "all" or a vector of gene names}
}
\value{
Returns a data.frame
}
\description{
Are there more genes containing >1 \emph{de novos} than expected?
}
\details{
See vignette (denovostats_intro) for more information.
}
\examples{
denovolyzeMultiHits(dnm.genes=autismDeNovos$gene,dnm.classes=autismDeNovos$dnmClass,nsamples=1078)

denovolyzeMultiHits(dnm.genes=autismDeNovos$gene,dnm.classes=autismDeNovos$dnmClass,include.class=c("syn","mis","non","splice","frameshift","lof","prot","all"),nsamples=1078)
}
\keyword{keywords}

