% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/denovolyzeR.doc.R
\docType{package}
\name{denovolyzeR}
\alias{denovolyzeR}
\alias{denovolyzeR-package}
\title{A package for the analysis of \emph{de novo} sequencing variants}
\description{
A package for the analysis of \emph{de novo} sequencing variants
}
\author{
James Ware \email{j.ware@imperial.ac.uk}
}
\references{
\url{http://github.com/jamesware/denovolyzeR}
}

