% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{FMRPgenes}
\alias{FMRPgenes}
\title{FMRP genes}
\format{A vector of (hgnc) gene identifiers}
\usage{
data(FMRPgenes)
}
\description{
842 genes found to interact with the fragile X mental retardation protein (FMRP)
}
\author{
James Ware
}
\references{
\url{http://dx.doi.org/10.1038/ng.3050}

\url{http://dx.doi.org/10.1016/j.cell.2011.06.013}
}
\keyword{datasets}

