% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FMRPgenes.doc.R
\docType{data}
\name{FMRPgenes}
\alias{FMRPgenes}
\title{FMRP genes}
\format{A vector of gene symbols}
\usage{
data(FMRPgenes)
}
\description{
842 genes found to interact with the fragile X mental retardation protein (FMRP)
}
\author{
James Ware
}
\references{
\url{http://dx.doi.org/10.1038/ng.3050}

\url{http://dx.doi.org/10.1016/j.cell.2011.06.013}
}
\keyword{datasets}

