% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMRPgenes.doc.R
\docType{data}
\name{fmrpGenes}
\alias{fmrpGenes}
\title{FMRP genes}
\format{A vector of gene symbols}
\description{
837 genes found to interact with the fragile X mental retardation protein (FMRP)
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4222185/}

\url{http://dx.doi.org/10.1016/j.cell.2011.06.013}
}
\keyword{datasets}

