% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseInput.R
\name{parseInput}
\alias{parseInput}
\title{Checks input for errors}
\usage{
parseInput(genes = genes, classes = genes, nsamples = nsamples,
  groupBy = groupBy, includeGenes = includeGenes,
  includeClasses = includeClasses, geneId = geneId, signifP = signifP,
  roundExpected = roundExpected)
}
\arguments{
\item{genes}{A vector of genes containing de novo variants.}

\item{classes}{A vector of classes of de novo variants.  Supported
classes are "syn", "mis", "mis_cons", "mis_dam", "mis_oth","non" or
"stoploss","startloss","splice", "frameshift" and "lof".}

\item{nsamples}{Number of individuals considered in de novo analysis.}

\item{groupBy}{Results can be tabulated by gene, or by variant class}

\item{includeGenes}{Genes to include in analysis. "all" or a vector of gene
names.}

\item{includeClasses}{Which variant classes are tabulated in output}

\item{geneId}{Gene identifier used. Currently only geneID. (refseqID may
work)}

\item{signifP}{Number of sig figs used to round p values in output.}

\item{roundExpected}{Number of decimal places used to roundExpected burdens
in output.}
}
\value{
tbc   , with or without error message
}
\description{
An internal function to check inputs
}

