% Generated by roxygen2 (4.0.2): do not edit by hand
\name{viewDeNovoProbs}
\alias{viewDeNovoProbs}
\title{Displays underlying \emph{de novo} probability tables}
\usage{
viewDeNovoProbs(format = "wide")
}
\description{
Tabulates probability of \emph{de novo} variant for each protein-coding variant class, for each gene.  Values are probability of a \emph{de novo} variant per chromosome per generation.  i.e. expected number of de novos for a given gene/class = \eqn{p * 2 * nsamples}.
}

